+++ scoserv/init_database.py
@@ -1,5 +1,5 @@
-"""Helper module to load an initial set of SCO models into the model
-repository. Expects the model definitions in a .json file.
+"""Helper module to create an empty database with an initial set of SCO models.
+Expects the model definitions in a .json file.
 """
 
 import json
@@ -10,7 +10,7 @@ from scodata.mongo import MongoDBFactory
 import scomodels
 
 
-def load_models(model_defs, mongo_collection, clear_collection=False):
+def load_models(model_defs, mongo_collection):
     """Load model defininitions into given mongo collection. The clear flag
     indicates whether existing models in the collection should be deleted.
 
@@ -20,12 +20,7 @@ def load_models(model_defs, mongo_collection, clear_collection=False):
         List of model definitions in Json-like format
     mongo_collection : MongoDB collection
         MongoDB collection where models are stored
-    clear_collection : boolean
-        If true, collection will be dropped before models are created
     """
-    # Drop collection if clear flag is set to True
-    if clear_collection:
-        mongo_collection.drop()
     # Create model registry
     registry = scomodels.DefaultModelRegistry(mongo_collection)
     for i in range(len(model_defs)):
@@ -41,8 +36,8 @@ if __name__ == '__main__':
     # Expect the configuration file as first and the model definition file as
     # second argument. An optional third argument contains the clear collection
     # flag (default: False)
-    if len(sys.argv) < 3 or len(sys.argv) > 4:
-        print 'Usage: <config-file> <model-defs-file> {<clear-collection-flag>}'
+    if len(sys.argv) != 3:
+        print 'Usage: <config-file> <model-defs-file>'
         sys.exit()
     # Read configuration file (YAML)
     with open(sys.argv[1], 'r') as f:
@@ -51,12 +46,7 @@ if __name__ == '__main__':
     # Read model definition file (JSON)
     with open(sys.argv[2], 'r') as f:
         models = json.load(f)
-    # Set clear collection flag if given
-    if len(sys.argv) == 4:
-        clear_collection = (sys.argv[3].upper() == 'TRUE')
-    else:
-        clear_collection = False
     # Get Mongo client factory
     mongo = MongoDBFactory(db_name=config['mongo.db'])
     # Load models
-    load_models(models, mongo.get_database().models, clear_collection=clear_collection)
+    load_models(models, mongo.get_database().models)